cambio

+++ ft_executor.c
@@ -6,7 +6,7 @@
 /*   By: msintas- <msintas-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 11:59:04 by msintas-          #+#    #+#             */
-/*   Updated: 2023/09/06 13:53:57 by msintas-         ###   ########.fr       */
+/*   Updated: 2023/09/09 23:59:10 by msintas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,9 +53,8 @@ void	ft_executor(t_content *cont)
             waitpid(pid, &status, 0);
             cont->global->err_stat = WEXITSTATUS(status);
         }
-        else
+        else if (cont[i].nfl == 2)
         {
-            //printf("infile es: %s\n", cont[i].infile);
             if (unlink(cont[i].infile) != 0)
             perror("unlink");
         }
